Fix bug: Deleting a project locally won't delete it from iCloud
The `removeFromCloud()` function now has a argument `deleteLocal: Bool` , which is used to control whether or not it should also delete the local data.
The `delete` function, used to delete projects locally, now checks whether a project also exists in the cloud. If it does it also delete the data in the cloud
The button to remove a project from the cloud was changed to pass `false` for the `deleteLocal` argument of `removeFromCloud()`